moved searchInput into searchStore
