+++ src/components/sidebar/LineItem.tsx
@@ -6,7 +6,6 @@ import { ILine, ILineRoute } from '../../models';
 import TransitTypeColorHelper from '../../util/transitTypeColorHelper';
 import Moment from 'react-moment';
 import * as s from './lineItem.scss';
-import lineStore from '../../stores/lineStore';
 import searchStore from '../../stores/searchStore';
 import { Location, History } from 'history';
 import LinkBuilder from '../../factories/linkBuilder';
@@ -58,7 +57,7 @@ class LineItem extends React.Component<ILineItemProps, ILineItemState> {
     public renderRoute(route: ILineRoute): any {
         const gotoUrl = (url:string) => () => {
             this.props.history.push(url);
-            lineStore.setSearchInput('');
+            searchStore.setSearchInput('');
             searchStore.removeAllSubLineItems();
         };
         return (
