+++ src/components/sidebar/RoutesView.tsx
@@ -4,26 +4,26 @@ import LineSearch from './LineSearch';
 import { Route, RouteComponentProps } from 'react-router';
 import RoutesList from './RoutesList';
 import { RouteStore } from '../../stores/routeStore';
-import { LineStore } from '../../stores/lineStore';
 import { SidebarStore } from '../../stores/sidebarStore';
 import { inject, observer } from 'mobx-react';
 import SearchResults from './SearchResults';
 import TransitToggleButtonBar from '../controls/TransitToggleButtonBar';
+import { SearchStore } from '../../stores/searchStore';
 
 interface ISidebarProps extends RouteComponentProps<any>{
     routeStore?: RouteStore;
-    lineStore?: LineStore;
+    searchStore?: SearchStore;
     sidebarStore?: SidebarStore;
 }
 
-@inject('routeStore', 'lineStore', 'sidebarStore')
+@inject('routeStore', 'searchStore', 'sidebarStore')
 @observer
 class RoutesView extends React.Component<ISidebarProps> {
     public render() {
         return (
             <div className={s.routesView}>
                 <LineSearch/>
-                { this.props.lineStore!.searchInput === '' ? (
+                { this.props.searchStore!.searchInput === '' ? (
                     <Route component={RoutesList} />
                 ) : (
                     <>
