+++ src/components/sidebar/Sidebar.tsx
@@ -4,26 +4,25 @@ import * as s from './sidebar.scss';
 import { Route, RouteComponentProps, Switch } from 'react-router-dom';
 import hslLogo from '../../assets/hsl-logo.png';
 import { RouteStore } from '../../stores/routeStore';
-import { LineStore } from '../../stores/lineStore';
+import searchStore from '../../stores/searchStore';
 import RoutesView from './RoutesView';
 import HomeView from './HomeView';
 
 interface ISidebarProps extends RouteComponentProps<any>{
     routeStore?: RouteStore;
-    lineStore?: LineStore;
 }
 
 interface ILinelistState {
     searchInput: string;
 }
 
-@inject('routeStore', 'lineStore')
+@inject('routeStore')
 @observer
 class Sidebar extends React.Component<ISidebarProps, ILinelistState> {
     public render(): any {
         const handleHeaderClick = () => {
             this.props.routeStore!.clearRoutes();
-            this.props.lineStore!.setSearchInput('');
+            searchStore.setSearchInput('');
             this.props.history.push('/');
         };
         return (
