+++ src/stores/lineStore.ts
@@ -4,11 +4,9 @@ import { ILine } from '../models';
 export class LineStore {
     @observable private _filters: string[];
     @observable private _allLines: ILine[];
-    @observable private _searchInput: string; // TODO: move this to searchStore
 
     constructor() {
         this._allLines = [];
-        this._searchInput = '';
     }
 
     @computed get filters(): string[] {
@@ -21,15 +19,6 @@ export class LineStore {
 
     @computed get allLines(): ILine[] {
         return this._allLines;
-    }
-
-    @computed get searchInput(): string {
-        return this._searchInput;
-    }
-
-    @action
-    public setSearchInput(input: string) {
-        this._searchInput = input;
     }
 
     @action
