+++ src/stores/searchStore.ts
@@ -1,13 +1,24 @@
 import { action, computed, observable } from 'mobx';
 
 export class SearchStore {
+    @observable private _searchInput: string;
     @observable private _subLineItems: {
         routePathId: string;
         routeId: string;
     }[];
 
     constructor() {
+        this._searchInput = '';
         this._subLineItems = [];
+    }
+
+    @computed get searchInput(): string {
+        return this._searchInput;
+    }
+
+    @action
+    public setSearchInput(input: string) {
+        this._searchInput = input;
     }
 
     @action
